// src/app/interfaces/user.ts
export interface User {
    nome: string;
    cpf: string;
    email: string;
    celular: string;
    profissao: string;
    dataNascimento: string;
    sexo: string;
    estadoCivil: string;
    cep: string;
    estado: string;
    cidade: string;
    bairro: string;
    rua: string;
    numero: string;
    complemento: string;
    membroIgrejaDoRecreio: string;
    cristao: string;
    eBatizado: string;
    dataBatismo: string;
    cidadeBatismo: string;
    igrejaBatismo: string;
    pastorBatismo: string;
    senha: string;
    confirmarSenha: string;
    aceitouTermos: boolean;
  }
  